Refine animations and links
This commit applies the smooth-scroll-middle animation to links in
index.html, such that the smooth scroll executes when the user clicks on
the "coming soon" button in the menu.

On top of that, changes have been made so that the "home" button in the
about page redirects to the root of the directory (/) instead of
index.html (/index.html).