Fix scrolling animation issue
This commit makes changes to the index HTML file to reflect new changes
that resolve the issue of no animations playing when "coming soon" is
clicked on.

Fix scrolling animation issue

This commit makes changes to the index HTML file to reflect new changes
that resolve the issue of no animations playing when "coming soon" is
clicked on.