💄 Introduce new design
This commit introduces a new design to the website. I thought that
having only black or white was too bland, so why not merge the two of
them together? This new version has is duo-tone with a dark colour and
white. For those devs out there, it kinda looks like GitHub, don't you
think?