⏪️ Revert TypeScript adoption
I'm still in the process of learning, so it'll be better for me to first
 understand concepts before adopting them. Adopting TypeScript was more
of a rash decision since I was introduced to its benefits, but I find
that there is no need for adopting it right now if I don't know the
difference between it and JavaScript entirely.

Once I've learned more about TypeScript, I can make a decision whether
to adopt it again or not.